splitting off ModelDiffractionVisualizations into separate file
